requirements.txt
notion-client==2.0.0
websockets==11.0.3
cryptography==41.0.1
Pillow==10.0.0
pynput==1.7.6
requests==2.31.0
python-dotenv==1.0.0 